pyproject.toml
[project]
name = "swp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.1",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.112",
    "jinja2>=3.1.6",
    "jose>=1.0.0",
    "jwt>=1.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
]
